+++ src/main/java/rsystems/commands/KarmaInterface.java
@@ -554,4 +554,5 @@ public class KarmaInterface extends ListenerAdapter {
         }
     }
 
+
 }