+++ src/main/java/rsystems/handlers/ReferenceLoader.java
@@ -3,57 +3,91 @@ package rsystems.handlers;
 import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
 import rsystems.HiveBot;
+import rsystems.adapters.ExtendedReference;
 import rsystems.adapters.Reference;
 
 import java.util.ArrayList;
 
 public class ReferenceLoader {
+    private static JSONFileHandler extendedReferenceFile = new JSONFileHandler("extendedReferenceData.json");
+    private static Object extendedReferenceCommands = extendedReferenceFile.getDatafileData();
+    private static JSONObject extendedReferenceData = (JSONObject) extendedReferenceCommands;
+
     private static JSONFileHandler referenceFile = new JSONFileHandler("referenceData.json");
     private static Object referenceCommands = referenceFile.getDatafileData();
     private static JSONObject referenceData = (JSONObject) referenceCommands;
 
     public ReferenceLoader(){
+        // Load extended references
+        extendedReferenceData.keySet().forEach(keyStr -> {
+            //Form the object into a JSONObject for processing
+            Object keyValue = extendedReferenceData.get(keyStr);
+            JSONObject parsedValue = (JSONObject) keyValue;
+
+            //Create a temporary Reference Object to hold the data
+            ExtendedReference tempExtendedReference = new ExtendedReference(
+                    keyStr.toString(),
+                    parsedValue.get("description").toString(),
+                    parsedValue.get("installation").toString()
+            );
+
+            try{
+                tempExtendedReference.setAliases(getArrayList(parsedValue,"alias"));
+            } catch(NullPointerException e){
+            }
+
+            tempExtendedReference.setLinks(getArrayList(parsedValue,"links"));
+            tempExtendedReference.setCategory(getArrayList(parsedValue,"category"));
+
+            //Add Reference Object into the references array
+            HiveBot.extendedReferences.add(tempExtendedReference);
+
+        });
+
+        // Load Simple references
         referenceData.keySet().forEach(keyStr -> {
             //Form the object into a JSONObject for processing
             Object keyValue = referenceData.get(keyStr);
             JSONObject parsedValue = (JSONObject) keyValue;
 
             //Create a temporary Reference Object to hold the data
-            Reference tempRef = new Reference(
+            Reference tempReference = new Reference(
                     keyStr.toString(),
-                    parsedValue.get("installation").toString(),
-                    parsedValue.get("description").toString().replace("{prefix}",HiveBot.prefix)
+                    parsedValue.get("description").toString()
             );
 
             try{
-                tempRef.setAlias(getArrayList(parsedValue,"alias"));
+                tempReference.setAliases(getArrayList(parsedValue,"alias"));
             } catch(NullPointerException e){
             }
-
-            tempRef.setLinks(getArrayList(parsedValue,"links"));
-            tempRef.setCategory(getArrayList(parsedValue,"category"));
+            tempReference.setCategory(getArrayList(parsedValue,"category"));
 
             //Add Reference Object into the references array
-            HiveBot.references.add(tempRef);
+            HiveBot.references.add(tempReference);
 
         });
     }
 
     public void updateData(){
+        extendedReferenceFile.loadDataFile();
+        extendedReferenceCommands = extendedReferenceFile.getDatafileData();
+        extendedReferenceData = (JSONObject) extendedReferenceCommands;
+
         referenceFile.loadDataFile();
         referenceCommands = referenceFile.getDatafileData();
         referenceData = (JSONObject) referenceCommands;
 
         // Clear all references from array to recreate them below.
+        HiveBot.extendedReferences.clear();
         HiveBot.references.clear();
 
-        referenceData.keySet().forEach(keyStr -> {
+        extendedReferenceData.keySet().forEach(keyStr -> {
             //Form the object into a JSONObject for processing
-            Object keyValue = referenceData.get(keyStr);
+            Object keyValue = extendedReferenceData.get(keyStr);
             JSONObject parsedValue = (JSONObject) keyValue;
 
                 //Create a temporary Reference Object to hold the data
-                Reference tempRef = new Reference(
+                ExtendedReference tempRef = new ExtendedReference(
                         keyStr.toString(),
                         parsedValue.get("installation").toString(),
                         parsedValue.get("description").toString().replace("{prefix}",HiveBot.prefix)
@@ -63,12 +97,35 @@ public class ReferenceLoader {
                 tempRef.setCategory(getArrayList(parsedValue,"category"));
 
                 try{
-                    tempRef.setAlias(getArrayList(parsedValue,"alias"));
+                    tempRef.setAliases(getArrayList(parsedValue,"alias"));
                 } catch(NullPointerException e){
                 }
 
                 //Add Reference Object into the references array
-                HiveBot.references.add(tempRef);
+                HiveBot.extendedReferences.add(tempRef);
+
+        });
+
+        // Load Simple references
+        referenceData.keySet().forEach(keyStr -> {
+            //Form the object into a JSONObject for processing
+            Object keyValue = referenceData.get(keyStr);
+            JSONObject parsedValue = (JSONObject) keyValue;
+
+            //Create a temporary Reference Object to hold the data
+            Reference tempReference = new Reference(
+                    keyStr.toString(),
+                    parsedValue.get("description").toString()
+            );
+
+            try{
+                tempReference.setAliases(getArrayList(parsedValue,"alias"));
+            } catch(NullPointerException e){
+            }
+            tempReference.setCategory(getArrayList(parsedValue,"category"));
+
+            //Add Reference Object into the references array
+            HiveBot.references.add(tempReference);
 
         });
     }